src/requirements.txt
django
djangorestframework
django-environ

